Overhaul the Infuser internals
This should make the Infuser perform better.
Change the Infuser's container to check if the recipe is valid and if not don't allow infusion.
Change it so you can't change the inventory of the Infuser during an Infusion without aborting the infusion.
Make infusions cost RF a tick.
Fix infusing with Force.
Add a sound for when the Infusion completes.
Make the Ender and Wing ability of the sword use durability instead of being free.
Make the Ender and Wing ability of the sword have a cooldown. (No flying)
Fix tooltips for certain Infusions. (Should now have a tooltip for all of them)
Fix some infusions not being possible even though the book says they are.
Add a recipe for the "Force and You" book.
Allow for Camo to be added to Armor pieces to not render them on the player.
Add more patchouli pages.